Moved the docs setup to the docs repository
Since a repository/project's Githuib Page is built into a subdirectory
with the same name, it made more sense to create a docs repository and
setup a regular mkdocs flow there.

This will reduce spam in the mail due to cancelled github page builds
and remove all workarounds I had in place to make mkdocs work from this
repository.

It will also give this repository a single responsibility, being the
location for the landing page of the project and the docs contain the
organisation's documentation